feat: optimize transport bar design with pattern buttons and improved controls
- Remove redundant position counter to free up space
- Add 4 quick-access pattern buttons with chaining support
- Replace tempo slider with precise up/down arrows (±0.1/1/10 BPM)
- Convert swing slider to compact rotary knob
- Add pattern length display
- Add CPU usage meter with color-coded warnings
- Add MIDI activity LED indicator
- Add panic button for emergency all-notes-off
- Reorganize layout with logical sections and visual dividers
- Implement new UI components: PatternButton, TempoArrows, CompactSwingKnob, PanicButton
- Fix timing and MIDI division issues in sequencer engine
- Add scale slot management system with 8 presets

+++ Tests/MidiEventGeneratorTests.cpp
@@ -80,7 +80,7 @@ private:
         events = generator.generateRatchetedEvents(72, 127, 1, 1000, 2);
         if (!events.empty() && events[0].message.isNoteOn())
         {
-            expectEquals(events[0].message.getVelocity(), 127, "Velocity should be 127");
+            expectEquals(static_cast<int>(events[0].message.getVelocity()), 127, "Velocity should be 127");
             expectEquals(events[0].channel, 2, "Channel should be 2");
         }
         
@@ -207,8 +207,8 @@ private:
             expectGreaterOrEqual(event.sampleOffset, 0, "Offset should be non-negative");
             if (event.message.isNoteOn())
             {
-                expectGreaterOrEqual(event.message.getVelocity(), 1, "Velocity should be >= 1");
-                expectLessOrEqual(event.message.getVelocity(), 127, "Velocity should be <= 127");
+                expectGreaterOrEqual(static_cast<int>(event.message.getVelocity()), 1, "Velocity should be >= 1");
+                expectLessOrEqual(static_cast<int>(event.message.getVelocity()), 127, "Velocity should be <= 127");
             }
         }
